requirement.txt
requests
json
tqdm
bs4
mysql-connector
